Reduce duplicated code from the protocol/private_id `stringify_id_map`
It makes more sense for it to remain a `protocol` method, rather than be imported from `common`, but I did just now move the bulk of the function into files.rs to reduce the amount duplicated code.  Do note for example, that your existing `save_id_map` and `print_id_map` are duplicates themselves, which I think is fine.