Upgrade the rand lib from rand0.7 to ran 0.8 (#63)
Summary:
Pull Request resolved: https://github.com/facebookresearch/Private-ID/pull/63

This is the to-do list when we migrate from github to fbcode. In order to migrate fast, we did not keep the lib version same as the internal version.

rand07 in the internal rust lib is temporary and can be removed in the future. We need to update the lib to rand 0.8 in rust to avoid build error when rand07 is removed.

{F753610665} {F753610678}

Reviewed By: rajprateek

Differential Revision: D37937653

fbshipit-source-id: 7c22cb21b0271047be0a8ead014a4aad30173f13