Add metrics logging to partner (#91)
Summary:
Pull Request resolved: https://github.com/facebookresearch/Private-ID/pull/91

# Context
We are adding metrics logging on multiple reasons.
1. To pass number of splits to Id Combiner as a workaround of [S3 API issue](https://fb.workplace.com/groups/pidmatchingxfn/posts/493743615908631).
2. To use union data size to calculate mutate number of MPC shards as mentioned in [the comment](https://fb.workplace.com/groups/164332244998024/posts/937649537666287/?comment_id=937655107665730&reply_comment_id=937696664328241)

To achieve above, we are enabling metrics logging on both single-key and multi-key PID protocols.

In addition, we found a small bug introduced in [#43](https://github.com/facebookresearch/Private-ID/pull/43) on publisher-side which is not logging metrics to a file when metrics path is not specified. We are fixing it in this diff.

# Description
The way it works is exactly the same as publisher-side metrics logging. We added a new arg `metric-path` in the executables.

1. When metrics path is specified, then metrics would be logged to the specified path.
2. When metrics path is not specified by output path is specified, then metrics would be logged to `{output path}_metrics` file.
3. When output path is not specified and stdout is used, metrics would be logged to stdout.

We are logging three metrics.
- partner_input_size
- publisher_input_size
- union_file_size

Reviewed By: shubho

Differential Revision: D39320561

fbshipit-source-id: c007d1c9aa2012cbb2768a3438869c5be09159a8